fully working version LET'S GOOOOOO

+++ compression_types.h
@@ -6,7 +6,7 @@
 
 
 typedef float COEFFICIENT_TYPE;
-typedef int16_t SAMPLE_TYPE; // will actually be uint16 with the iadc, but for testing, it's easier to have it float
+typedef float SAMPLE_TYPE; // will actually be uint16 with the iadc, but for testing, it's easier to have it float
 #define SIGNAL_LENGTH 32
 #define COEFFICIENTS_LENGTH 42
